ADD :: auth부분 쿼리 캡슐화
