+++ src/features/auth/login/login.props.ts
@@ -1,0 +1,5 @@
+import { LoginType } from 'types/auth/auth.type';
+
+export interface LoginFeatureType {
+  loginData: LoginType;
+}
