+++ src/features/auth/school/school.props.ts
@@ -1,0 +1,3 @@
+export interface SchoolFeatureType {
+  searchWord: string;
+}
